Adds start() and resume() to TestHelper.initBaseActivity()
This is actually a significant improvement since it allows fragment transactions invoked using BaseActivity's fragment manger to actually be committed in the test harness. This was a source of many issues when spawning a new fragment as part of a test.

Also overrides getMap() in TestBaseActivity to return a TestMap instance since this is called as part of the  initialization of MapFragment which is declared in xml and occurs when performStart() is invoked on BaseActivity.